Add avahi implementation of bonjour
This solves several problems. First, if there is more that one bonjour implementatinon running at the same time, it confuses both implementations because they are using the same multicast address/port.

Also, the pure js implementation seems to only bind to the default network interface, so it doesn't work so well when there are multiple network interfaces.