Add moreComing flag to serivce added event
This adds a moreComing flag to the dnssd.Service.on('added', ...) event. This comes from the Apple implementation. It is used to let clients know not to update the UI yet becuase there are more changes on the way. It is implemented as best as possible in the other backends.