Make it easy for users to enter device info
This adds a new entry "I don't see my device..." to the quick pick menu for picking a device.

This allows the user to enter a name and IP address for the device. The information is saved in the global settings so that it only has to be entered once.