Present useful error when part of path given to mkdir_p exists but isn't a directory (#39)
Currently, if the `mkdir_p` method is given a path where a segment of it exists but isn't a directory, the sftp server returns error code 4 ("failure") with just the message "failure". With this change, we explain what went wrong.

Note that I also reversed the order of slash additions, so that `mkdir` is called without a trailing slash. This is because the sftp `stat` call returns that the file doesn't exist if the trailing slash is included; both `stat` and `mkdir` operate as expected without it, with the added benefit that `stat` succeeds instead of claiming that the file doesn't exist. This way, we can handle it intelligently.