Update to more modern typescript
This all started with just trying to update tasks.json, launch.json and a few other files to the current yo code template. This required upgrading typescript which in turn required making a bunch of changes to fix the more strict type checking.

Theoretically, there are no functional changes. A quick manual test shows everything to still be working (on Linux).