Fix duplicate devices listed in quick-pick on Windows
The bonjour backend used on Windows was not properly filtering duplicate notifications on the same network adapter properly, resulting in each device being listed more than once. The logic for ignoring duplicates is similar to the macOS (dnssd) backend.

Also, there was a bug introduced by switching to === that caused the network adapter name to not be shown in the quick-pick on Windows too (because bClient.iface was string instead of number). This was fixed as a side-effect of fixing the duplicates.