New "pause" button on debugger that sends SIGINT to remote process
Since fixing the restart button, the pause button is always enabled.
There doesn't seem to be a way to disable it, so let's make it do
something useful. For example, if a python program is run with
PYTHONINSPECT=1 environment variable, then the pause button
will interrupt the program an open a REPL for inspecting the current
program state.

Know bug: pause button does not work after restarting with the restart
button. This is caused by <https://github.com/microsoft/vscode/issues/88784>.