Fix race condition when browsing for connected devices
Event handlers were connected after browsing started, so device added
events could be missed, causing devices to not be listed in the
quickpick menu.

This is fixed by adding new start and stop methods to the browser
classes on each backend. This way the object can be created, event
listeners attached, then the browser can be started.