automatically convert files with #! and CRLF
It is a common mistake for Windows users to not notice the line endings
of the file. The default on Windows is to use CRLF which causes files
with a #! to fail to execute when run on the EV3. This automatically
converts the file when downloading so that users don't have to worry
about it.